requirements.txt
# Build dependencies - pinned for reproducibility
PyYAML==6.0.1
Jinja2==3.1.4
Markdown==3.7
beautifulsoup4==4.12.3
requests==2.32.3

# Optional: LLM integration (Anthropic)
anthropic==0.39.0
